agent-poller, test: Extract AsyncAgentStore tests to new file
Extract AsyncAgentStore unit tests to a dedicated agent poller test file.
