+++ pkg/virt-launcher/virtwrap/agent-poller/BUILD.bazel
@@ -21,6 +21,7 @@ go_test(
     srcs = [
         "agent_parser_test.go",
         "agent_poller_suite_test.go",
+        "agent_poller_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
