+++ pkg/virt-launcher/virtwrap/agent-poller/agent_parser_test.go
@@ -372,61 +372,5 @@ var _ = Describe("Qemu agent poller", func() {
 			Expect(err).ToNot(HaveOccurred(), "users should be parsed normally")
 			Expect(users).To(Equal(expectedUsers))
 		})
-
-		Context("with AsyncAgentStore", func() {
-
-			It("should store and load the data", func() {
-				var agentStore = NewAsyncAgentStore()
-				agentVersion := "4.1"
-				agentStore.Store(GET_AGENT, agentVersion)
-				agent := agentStore.GetGA()
-
-				Expect(agent).To(Equal(agentVersion))
-			})
-
-			It("should fire an event for new sysinfo data", func() {
-				var agentStore = NewAsyncAgentStore()
-
-				fakeInfo := api.GuestOSInfo{
-					Name:          "TestGuestOSName",
-					KernelRelease: "1.1.0-Generic",
-					Version:       "1.0.0",
-					PrettyName:    "TestGuestOSName 1.0.0",
-					VersionId:     "1.0.0",
-					KernelVersion: "1.1.0",
-					Machine:       "x86_64",
-					Id:            "testguestos",
-				}
-				agentStore.Store(GET_OSINFO, fakeInfo)
-
-				Expect(agentStore.AgentUpdated).To(Receive(Equal(AgentUpdatedEvent{
-					Type:       GET_OSINFO,
-					DomainInfo: api.DomainGuestInfo{OSInfo: &fakeInfo},
-				})))
-			})
-
-			It("should not fire an event for the same sysinfo data", func() {
-				var agentStore = NewAsyncAgentStore()
-				fakeInfo := api.GuestOSInfo{
-					Name:          "TestGuestOSName",
-					KernelRelease: "1.1.0-Generic",
-					Version:       "1.0.0",
-					PrettyName:    "TestGuestOSName 1.0.0",
-					VersionId:     "1.0.0",
-					KernelVersion: "1.1.0",
-					Machine:       "x86_64",
-					Id:            "testguestos",
-				}
-
-				agentStore.Store(GET_OSINFO, fakeInfo)
-				Expect(agentStore.AgentUpdated).To(Receive(Equal(AgentUpdatedEvent{
-					Type:       GET_OSINFO,
-					DomainInfo: api.DomainGuestInfo{OSInfo: &fakeInfo},
-				})))
-
-				agentStore.Store(GET_OSINFO, fakeInfo)
-				Expect(agentStore.AgentUpdated).ToNot(Receive())
-			})
-		})
 	})
 })
